requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
ollama
